requirements.txt
pynput==1.7.6
RPi.GPIO==0.7.1
apds9960==0.2
smbus==1.1.post2
